requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
matplotlib==3.10.1
Pillow==10.3.0
mlflow==2.20.4
joblib==1.4.2
scikit-surprise==1.1.4
gdown==5.2.0
